server/app/requirements.txt
asgiref==3.4.0
autopep8==1.5.7
bandit==1.7.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
flake8==3.9.2
gitdb==4.0.7
GitPython==3.1.18
hashids==1.3.1
idna==2.10
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
oauthlib==3.1.1
pbr==5.6.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.3.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
